Stop allowing submissions to run on a subset of testcases
Reverts b6fb6ec324f705f35feb7151522028b9608bf9c1
